Cleaned up Startup for initializing the DB.
