+++ .idea/misc.xml
@@ -8,7 +8,7 @@
         <entry key="..\:/Users/gdbar/AndroidStudioProjects/WordSpell/app/src/main/res/layout/activity_main.xml" value="0.14635416666666667" />
         <entry key="..\:/Users/gdbar/AndroidStudioProjects/WordSpell/app/src/main/res/layout/activity_main_menu.xml" value="0.1743659420289855" />
         <entry key="..\:/Users/gdbar/AndroidStudioProjects/WordSpell/app/src/main/res/layout/activity_results.xml" value="0.2" />
-        <entry key="..\:/Users/gdbar/AndroidStudioProjects/WordSpell/app/src/main/res/layout/activity_startup.xml" value="0.536" />
+        <entry key="..\:/Users/gdbar/AndroidStudioProjects/WordSpell/app/src/main/res/layout/activity_startup.xml" value="0.2" />
         <entry key="..\:/Users/gdbar/AndroidStudioProjects/WordSpell/app/src/main/res/layout/activity_word_spell.xml" value="0.16" />
       </map>
     </option>
