+++ app/src/main/java/com/example/wordspell/Startup.kt
@@ -1,13 +1,10 @@
 package com.example.wordspell
 
-import android.content.Context
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.widget.TextView
 import android.widget.Button
 import android.content.Intent
-import android.util.Log
-
+import android.widget.TextView
 
 
 class Startup : AppCompatActivity() {
@@ -16,8 +13,11 @@ class Startup : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_startup)
 
-        // Load the user's data
-        loadData()
+        // Setup DB
+        val db = DBHelper(this, null)
+        // Initialize DB and load data
+        loadData(db)
+        crossCheckList(db)
 
 
         // Gets the button and listens for the click to go to the next Activity
@@ -43,50 +43,56 @@ class Startup : AppCompatActivity() {
 
         val button4 = findViewById<Button>(R.id.button4)
         button4.setOnClickListener {
-            val intent = Intent(this@Startup, DBTest::class.java)
+            val intent = Intent(this@Startup, HighScore::class.java)
             startActivity(intent)
             finish()
         }
+
+        val testView = findViewById<TextView>(R.id.textView2)
+        Global.setScore("apple", 0, this)
+        testView.text = Global.getScore("apple").toString()
     }
 
-    fun loadData() {
-        // Current implementation is hard coded.
-        // This will be altered according to database.
-        val wordList = arrayOf(
-            "apple",
-            "bath",
-            "boat",
-            "child",
-            "clock",
-            "dress",
-            "eight",
-            "eleven",
-            "gray",
-            "happy",
-            "house",
-            "lock",
-            "lunch",
-            "night",
-            "pool",
-            "school",
-            "sea",
-            "shape",
-            "sky",
-            "slide",
-            "stone",
-            "truck",
-            "tune",
-            "winter"
-        )
-        val scoreList = 0
-        val scoreDateTime  = "2022.02.16 at 2:27:00"
 
-        for (i in wordList.indices) {
-            val wordDataObject = WordData(wordList[i], scoreList, scoreDateTime)
-            Global.setCurrentDataList(wordDataObject)
+    fun loadData(db : DBHelper) {
+        // Clears the Global local data for refresh from db
+        Global.clearWordData()
+        Global.clearWordList()
+
+        // Get the DB data
+        val readList = db.readData()
+
+        // Put the contents into the Global local data
+        Global.setGlobalData(readList)
+        Global.generateWordList()
+    }
+
+    fun crossCheckList(db : DBHelper) {
+        // Get the hard coded word list
+        val allWordsList = WordList.allWordsList
+
+        // Check if it exists in the db loaded data
+        for(refWord in allWordsList) {
+            val check = Global.wordList.find { word -> word == refWord}
+            // If the word doesn't exist, add it to the database with parameters
+            if(check == null) {
+                // Set the parameters
+                val word = refWord
+                val score = 0
+                val date = "2022.02.16 at 2:27:00"
+
+                // Create the WordData object
+                val newWord = WordData(word, score, date)
+
+                // Put into the db
+                db.insertData(newWord)
+
+                // Add to Global
+                Global.setCurrentDataList(newWord)
+                Global.addOneToWordList(newWord.word)
+            }
         }
 
-        Global.generateWordList()
     }
 
     fun goToMainMenu() {
